refactor: land issue stack 6 through 10
