+++ src/research_lab/engine.py
@@ -4,8 +4,9 @@ from collections import Counter
 import os
 from pathlib import Path
 
+from research_lab.identity import candidates_match
 from research_lab.llm import LlmClient, LlmError, rerank_candidates_with_llm, summarize_candidates_with_llm
-from research_lab.models import PaperCandidate, QueryRecord, ResearchBrief, RunArtifacts
+from research_lab.models import EnrichedCandidate, PaperCandidate, QueryRecord, ResearchBrief, RetrievalCandidate, RunArtifacts, ScoredCandidate
 from research_lab.planner import build_expansion_queries, build_seed_queries
 from research_lab.rank import dedupe_candidates, extract_evidence_sentences, rank_candidates
 from research_lab.report import write_run_files
@@ -34,7 +35,7 @@ def execute_run(
     semantic_scholar_api_key = os.getenv("SEMANTIC_SCHOLAR_API_KEY", "").strip()
     all_queries: list[QueryRecord] = []
     seen_query_strings: set[str] = set()
-    pool: list[PaperCandidate] = []
+    pool: list[PaperCandidate | RetrievalCandidate] = []
     warnings: list[str] = []
     source_state = {
         "arxiv_enabled": True,
@@ -82,16 +83,19 @@ def execute_run(
         pool.extend(expanded_pool)
         ranked = rank_candidates(dedupe_candidates(pool), brief)
 
-    _enrich_top_candidates(ranked, brief, client, warnings)
-    ranked = rank_candidates(dedupe_candidates(pool), brief)
-    synthesis = _apply_llm_layer(ranked, brief, warnings)
+    enriched = _enrich_top_candidates(ranked[: brief.full_text_top_n], brief, client, warnings)
+    pool.extend(candidate.to_paper_candidate() for candidate in enriched)
+    deduped_pool = dedupe_candidates(pool)
+    ranked = rank_candidates(deduped_pool, brief)
+    final_ranked = [EnrichedCandidate.from_paper_candidate(candidate) for candidate in _merge_scored_candidates(ranked, deduped_pool)]
+    synthesis = _apply_llm_layer(final_ranked, brief, warnings)
 
     artifacts = RunArtifacts.create(
         run_id=run_id,
         run_dir=str(run_dir),
         brief=brief,
         queries=all_queries,
-        candidates=ranked,
+        candidates=[candidate.to_paper_candidate() for candidate in final_ranked],
         program_text=program_text,
         warnings=sorted(set(warnings)),
         synthesis=synthesis,
@@ -108,8 +112,8 @@ def _search_all_sources(
     client: HttpClient,
     warnings: list[str],
     source_state: dict[str, object],
-) -> list[PaperCandidate]:
-    collected: list[PaperCandidate] = []
+) -> list[RetrievalCandidate]:
+    collected: list[RetrievalCandidate] = []
     if _should_use_arxiv(query, source_state):
         try:
             source_state["arxiv_requests_remaining"] -= 1
@@ -219,12 +223,30 @@ def _add_query(query: QueryRecord, all_queries: list[QueryRecord], seen: set[str
     return True
 
 
-def _top_authors(candidates: list[PaperCandidate]) -> list[str]:
+def _top_authors(candidates: list[ScoredCandidate]) -> list[str]:
     counts = Counter(author for candidate in candidates for author in candidate.authors[:3] if author)
     return [author for author, _ in counts.most_common(3)]
 
 
-def _expansion_seed_candidates(ranked: list[PaperCandidate], brief: ResearchBrief) -> list[PaperCandidate]:
+def _merge_scored_candidates(ranked: list[ScoredCandidate], pool: list[PaperCandidate]) -> list[PaperCandidate]:
+    merged: list[PaperCandidate] = []
+    used_indices: set[int] = set()
+    for scored_candidate in ranked:
+        for index, candidate in enumerate(pool):
+            if index in used_indices:
+                continue
+            if not candidates_match(scored_candidate.to_paper_candidate(), candidate):
+                continue
+            candidate.score = scored_candidate.score
+            candidate.reasons = list(scored_candidate.reasons)
+            candidate.flags = list(scored_candidate.flags)
+            merged.append(candidate)
+            used_indices.add(index)
+            break
+    return merged
+
+
+def _expansion_seed_candidates(ranked: list[ScoredCandidate], brief: ResearchBrief) -> list[ScoredCandidate]:
     required_hits = 0
     if brief.must_include:
         required_hits = 1 if len(brief.must_include) <= 2 else 2
@@ -235,7 +257,7 @@ def _expansion_seed_candidates(ranked: list[PaperCandidate], brief: ResearchBrie
         if candidate.document_kind == "paper"
         and candidate.score >= 0.45
         and _must_include_hits(candidate, brief) >= required_hits
-        and not any("drift" in reason or "weak title alignment" == reason for reason in candidate.reasons)
+        and not any(flag in {"drift", "weak_title"} for flag in candidate.flags)
     ]
     if filtered:
         return filtered[:3]
@@ -246,37 +268,40 @@ def _expansion_seed_candidates(ranked: list[PaperCandidate], brief: ResearchBrie
     return ranked[:3]
 
 
-def _must_include_hits(candidate: PaperCandidate, brief: ResearchBrief) -> int:
-    searchable_text = f"{candidate.title} {candidate.abstract} {candidate.snippet} {candidate.full_text}".lower()
+def _must_include_hits(candidate: ScoredCandidate, brief: ResearchBrief) -> int:
+    searchable_text = f"{candidate.title} {candidate.abstract} {candidate.snippet}".lower()
     return sum(1 for term in brief.must_include if term.lower() in searchable_text)
 
 
 def _enrich_top_candidates(
-    ranked: list[PaperCandidate],
+    ranked: list[ScoredCandidate],
     brief: ResearchBrief,
     client: HttpClient,
     warnings: list[str],
-) -> None:
-    for candidate in ranked[: brief.full_text_top_n]:
-        if candidate.full_text:
-            continue
+) -> list[EnrichedCandidate]:
+    enriched: list[EnrichedCandidate] = []
+    for candidate in ranked:
         try:
             result = fetch_candidate_full_text(candidate, client)
         except SourceError as exc:
             warnings.append(str(exc))
             continue
-        candidate.access_status = result.access_status
-        candidate.access_url = result.access_url
+        enriched_candidate = EnrichedCandidate.from_paper_candidate(candidate.to_paper_candidate())
+        enriched_candidate.access_status = result.access_status
+        enriched_candidate.access_url = result.access_url
         if not result.text:
             if result.access_status == "unreadable":
-                warnings.append(f"no readable content found for {candidate.title}")
+                warnings.append(f"no readable content found for {enriched_candidate.title}")
+            enriched.append(enriched_candidate)
             continue
-        candidate.full_text = result.text
-        candidate.full_text_source = result.source
-        candidate.evidence = extract_evidence_sentences(result.text, brief)
+        enriched_candidate.full_text = result.text
+        enriched_candidate.full_text_source = result.source
+        enriched_candidate.evidence = extract_evidence_sentences(result.text, brief)
+        enriched.append(enriched_candidate)
+    return enriched
 
 
-def _apply_llm_layer(ranked: list[PaperCandidate], brief: ResearchBrief, warnings: list[str]) -> str:
+def _apply_llm_layer(ranked: list[EnrichedCandidate], brief: ResearchBrief, warnings: list[str]) -> str:
     client = LlmClient.from_env()
     if client is None:
         return ""
